error scope utility now scopes over all scope types
